[QoL] moved puzzle generation from BossFactory to new PuzzleGenerator
