+++ qrogue/game/logic/__init__.py
@@ -3,6 +3,7 @@ import qrogue.game.logic.actors
 import qrogue.game.logic.base
 import qrogue.game.logic.collectibles
 from qrogue.game.logic.message import Message
+from qrogue.game.logic.puzzle_generator import PuzzleGenerator
 
 # importing
 # +util
